Remove dependency of Selenium
