+++ examples/web_ui/caqui/constants.py
