+++ examples/web_ui/caqui/synchronous/home.py
@@ -4,7 +4,7 @@
 # Visit: https://github.com/douglasdcm/guara
 
 from guara.transaction import AbstractTransaction
-from examples.web_ui.caqui_async.constants import MAX_INDEX
+from examples.web_ui.caqui.constants import MAX_INDEX
 
 
 class GetNthLink(AbstractTransaction):
