+++ examples/web_ui/playwright/synchronous/pages/home.py
