+++ examples/web_ui/playwright/synchronous/pages/setup.py
