+++ examples/web_ui/playwright/synchronous/test_web_page.py
@@ -5,7 +5,7 @@
 
 from pytest import fixture, mark
 from playwright.sync_api import Page
-from examples.web_ui.playwright.pages import home, setup, getting_started
+from examples.web_ui.playwright.synchronous.pages import home, setup, getting_started
 from guara.transaction import Application
 from guara import it
