+++ examples/web_ui/selenium/simple/test_local_page.py
@@ -6,15 +6,14 @@
 from pathlib import Path
 from random import randrange
 from selenium import webdriver
-from examples.web_ui.selenium.simple.local_page import home
+from examples.web_ui.selenium.simple import home, setup
 from guara.transaction import Application
 from guara import it
-from guara import setup
 
 
 class TestLocalPage:
     def setup_method(self, method):
-        file_path = Path(__file__).parent.parent.parent.parent.resolve()
+        file_path = Path(__file__).parent.parent.parent.resolve()
         options = webdriver.ChromeOptions()
         options.add_argument("--headless=new")
         self._app = Application(webdriver.Chrome(options=options))
@@ -29,7 +28,7 @@ class TestLocalPage:
     def teardown_method(self, method):
         self._app.at(setup.CloseApp)
 
-    def test_local_page(self):
+    def test_local_page_selenium(self):
         text = ["cheese", "selenium", "test", "bla", "foo"]
         text = text[randrange(len(text))]
         self._app.at(home.SubmitText, text=text).asserts(it.IsEqualTo, f"It works! {text}!")
