+++ examples/web_ui/undetected_chromedriver/setup.py
@@ -1,3 +1,8 @@
+# Copyright (C) 2025 Guara - All Rights Reserved
+# You may use, distribute and modify this code under the
+# terms of the MIT license.
+# Visit: https://github.com/douglasdcm/guara
+
 from guara.transaction import AbstractTransaction
