+++ examples/web_ui/undetected_chromedriver/test_undetected_chromedriver.py
@@ -1,3 +1,8 @@
+# Copyright (C) 2025 Guara - All Rights Reserved
+# You may use, distribute and modify this code under the
+# terms of the MIT license.
+# Visit: https://github.com/douglasdcm/guara
+
 import pytest
 import undetected_chromedriver as uc
 from guara.transaction import Application
